requirements.txt
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
pillow==8.4.0
progress==1.6
pyyaml==6.0
torch-dct==0.1.5
